DIABLO-312 DIABLO-313 Move remaining email types into queue
