+++ tests/test_jobs/test_kaltura_job.py
@@ -24,6 +24,7 @@ ENHANCEMENTS, OR MODIFICATIONS.
 """
 from diablo import std_commit
 from diablo.jobs.kaltura_job import KalturaJob
+from diablo.jobs.queued_emails_job import QueuedEmailsJob
 from diablo.models.approval import Approval
 from diablo.models.room import Room
 from diablo.models.scheduled import Scheduled
@@ -110,6 +111,7 @@ class TestKalturaJob:
             assert scheduled.term_id == term_id
 
             # Verify emails sent
+            QueuedEmailsJob(app.app_context).run()
             emails_sent = _get_emails_sent()
             assert len(emails_sent) == email_count + 2
             assert emails_sent[-1].recipient_uids + emails_sent[-2].recipient_uids == ['10009', '10010']
